Enhance configuration migration functionality and SQL schema
- Added automatic null analysis mode for Java compilation in VSCode settings.
- Updated ConfigMigrationMapper to use BigInteger for ID parameters, improving compatibility with larger IDs.
- Introduced a new insert method in ConfigVersionMapper for adding configuration version records.
- Enhanced DataMigrationUtil to safely handle ID retrieval as BigInteger and added error logging for migration failures.
- Updated application.yml to disable migration by default.
- Modified SQL migration script to backup and alter tables for version control, including new fields and indexes.
- Improved test cases in DataMigrationUtilTest to validate initial states and batch processing.

These changes strengthen the migration process, ensuring better data integrity and handling during configuration updates.

+++ src/main/java/com/example/mapper/ConfigMigrationMapper.java
@@ -1,6 +1,7 @@
 package com.example.mapper;
 
 import org.apache.ibatis.annotations.*;
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Map;
 
@@ -14,7 +15,7 @@ public interface ConfigMigrationMapper {
     void updateVersionId(
         @Param("tableName") String tableName,
         @Param("versionId") String versionId,
-        @Param("id") Long id
+        @Param("id") BigInteger id
     );
     
     @Select("SELECT COUNT(*) FROM ${tableName} WHERE version_id IS NULL")
